Revert "entity files have been renamed"
This reverts commit a9f19544cd8246b472d6c3194a18f758c4da48a7.
