+++ server/entity/Point.py
